+++ jena-integration-tests/src/test/java/org/apache/jena/test/rdflink/TestRDFLinkFusekiBinary.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.jena.integration;
+package org.apache.jena.test.rdflink;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
