+++ jena-integration-tests/src/test/java/org/apache/jena/test/rdflink/TestRDFLinkLocalTDB.java
@@ -16,23 +16,24 @@
  * limitations under the License.
  */
 
-package org.apache.jena.integration;
+package org.apache.jena.test.rdflink;
 
+import org.apache.jena.rdfconnection.Isolation;
+import org.apache.jena.rdflink.AbstractTestRDFLink;
 import org.apache.jena.rdflink.RDFLink;
 import org.apache.jena.rdflink.RDFLinkFactory;
-import org.apache.jena.rdflink.RDFLinkFuseki;
+import org.apache.jena.sparql.core.DatasetGraph;
+import org.apache.jena.tdb.TDBFactory ;
 
-public class TestRDFLinkFuseki extends TestRDFLinkHTTP {
+public class TestRDFLinkLocalTDB extends AbstractTestRDFLink {
+
+    @Override
+    protected boolean supportsAbort() { return true ; }
+
     @Override
     protected RDFLink link() {
-        return RDFLinkFactory.connectFuseki("http://localhost:"+PORT+"/ds");
-    }
-
-    @Override
-    protected boolean defaultToCheckQueries() { return false; }
-
-    @Override
-    protected RDFLink link(boolean parseCheckSPARQL) {
-        return RDFLinkFuseki.service("http://localhost:"+PORT+"/ds").parseCheckSPARQL(parseCheckSPARQL).build();
+        DatasetGraph dsg = TDBFactory.createDatasetGraph() ;
+        return RDFLinkFactory.connect(dsg, Isolation.COPY) ;
     }
 }
+
